Colored: Don't print anything for empty arrays
In some places, we have a `Colored` struct with an empty value,
for example when the `path` is empty. This is useful to simplify
the call site, as a ternary must have a common type,
and using the same number of arguments for a `format` call avoids branching.
However, it inserts (invisible) color markers in the output,
which then makes comparing the result cumbersome.
The simple workaround is to omit the colors if the value is empty,
as they will be invisible anyways.