Fix #2: Make usage of colors in error message configurable
In order for this to be possible, the caller needs to decide whether
colors are wanted or not, which is done via a custom formatting string.
This means that we need to encode the format into the Exception
(with custom Exception types) instead of doing it at the `throw` site.

This was achieved by the previous commits, introducing an Exception
type for each kind of error encountered while parsing the config file.
This, combined with a custom format string, achieve the desired outcome.
This commit "fixes" the issue, but simply remove the leftover 'enforce'
method, which is no longer used (as it assumed ahead of time formatting),
and related utilities.