Add support for `fromYAML` hook
The converter approach is useful for fields,
but it falls short in the face of type composition,
e.g. when the field is an array.
Using a `fromYAML` hook turned out to be a superior approach,
as proven while porting DUB to use Configy.

As `fromYAML` is a more generic way to do `Converter`,
the latter are no longer documented in `configy.Read`.