Improve support for static arrays and non-CTFEable opEquals
The original intent of this commit was to avoid calling `opEquals`
on types that we were evaluating for optionality, as a simple `is`
comparison is enough. However, `is` was not used previously because
it triggers a deprecation in D: Doing `a is b` when `a` and `b`
are static array would compare their `.ptr`.

Re-writing `isOptional` and adding test yield some issues with static
array support, including compilation error. This is now working and
the library will throw a proper error if the length do not match.